refactor(docs): extract _inject_counter_values into a Sphinx extension
Moves _inject_counter_values (event: source-read) from conf.py into
docs/source/_extensions/counter_values.py. Substitutes placeholder tokens
like |datasets_total|, |subjects_total|, etc. into the dataset_summary
RST source so the counters render correctly even with JavaScript disabled.
Helpers moved alongside: _split_tokens, _compute_dataset_counter_defaults,
_format_counter, _DATASET_COUNTER_DEFAULTS, _DATASET_COUNTER_PLACEHOLDERS
(sole callers). conf.py shrinks from 4076 to 3971 LOC.

Part of C1 in docs_pipeline_architecture_review.md
